Update validation and sanitization functions
