+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@gcforms/core",
-  "version": "2.2.23",
+  "version": "2.2.24",
   "author": "Canadian Digital Service",
   "license": "MIT",
   "publishConfig": {
